Extract native review pane
