+++ docs/DECISIONS.md
@@ -111,6 +111,7 @@
 - The native model picker is a separate SwiftUI control instead of a nested top-bar private block. The picker owns search focus, mode selection, category sections, row expansion, favorite toggles, and model selection together, with 40 pt hit targets, shared `0.96` press feedback, and middle truncation for long provider/model metadata.
 - The native top bar is a separate SwiftUI control. Workspace layout owns where the chrome sits, while `QuillCodeTopBarView` owns thread identity, model/mode access, runtime status, and the bounded utility overflow. The ellipsis uses the shared 40 pt hit target and `0.96` press feedback so the quiet Codex-like chrome still feels tactile.
 - The native sidebar is a separate SwiftUI control. The workspace shell decides width and placement, while `QuillCodeSidebarView` owns primary navigation, thread sections, bulk selection, project rows, and the compact Tools/Settings footer. Sidebar commands still route through shared `WorkspaceCommandSurface` IDs, and high-frequency sidebar controls follow the same 40 pt hit target and `0.96` press feedback contract as other chrome.
+- The native review pane is a separate SwiftUI control. `QuillCodeReviewPaneView` owns review summary, file rows, hunk rows, inline comments, range notes, and review action buttons, while the workspace shell only decides when the review surface appears. Review controls use the shared 40 pt hit target and `0.96` press feedback contract so diff-review actions stay compact but tactile.
 - Automations and Activity are separate surfaces. Activity is the current thread's live execution log: task plan, tools, sources, artifacts, and latest answer. Automations is for persisted or scheduled future work: thread follow-ups, workspace schedules, monitors, and later subagents. Planned workflow rows remain visible only when no configured automations exist, so the Automations pane can teach the shape of upcoming runtime work without inventing fake jobs.
 - Automation records persist as a single `~/.quillcode/automations.json` array of `QuillAutomation` values. The workspace model loads those records at bootstrap and projects them into the Automations pane, falling back to planned workflow templates only when no jobs exist. The scheduler/runner is intentionally separate from this store so heartbeat, cron, monitor, and notification execution can be added without changing the visible surface contract.
 - Automation management is command-driven. `automation-create-thread-follow-up` creates a persisted follow-up for the selected thread, `automation-create-workspace-schedule` creates a persisted schedule for the selected project, `automation-run:<id>` manually wakes runnable jobs, and row actions use `automation-pause:<id>`, `automation-resume:<id>`, and `automation-delete:<id>`. SwiftUI, static HTML, and Playwright render those command IDs from the same surface records instead of owning separate automation reducers.
